# Conflicts:
#	Magic/src/main/java/com/elmakers/mine/bukkit/arena/Arena.java
#	Magic/src/main/java/com/elmakers/mine/bukkit/arena/ArenaStage.java

+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/MageModifier.java
@@ -80,7 +80,7 @@ public class MageModifier extends BaseMageModifier implements Comparable<MageMod
     @Override
     public int getTimeRemaining() {
         long now = System.currentTimeMillis();
-        return (int)(getLong("last_reset") + (long)getDuration() - now);
+        return (int)(getLong("last_reset") + getDuration() - now);
     }
 
     @Override
